실시간 위치 기록하여 firestore에 업데이트 기능 추가

+++ app/src/main/java/com/example/gpsdiary/spot/LocationClass.kt
@@ -1,4 +1,4 @@
 package com.example.gpsdiary.spot
 
-class LocationClass(var latitude: Float, var longitude: Float) {
+class LocationClass(var latitude: Double, var longitude: Double) {
 }
